Solved AoC2020 Day 11 Part 1

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day11/Day11Test.kt
@@ -10,7 +10,7 @@ class Day11Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         SeatingSystem(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
@@ -22,7 +22,7 @@ class Day11Test {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day11/day11part1sample.txt", 37L)
+                Arguments.of("de/havox_design/aoc2020/day11/day11part1sample.txt", 37)
             )
 
         @JvmStatic
